requirements.txt
beautifulsoup4==4.11.1
discord.py==1.7.3
requests==2.27.1
